# code/java-client/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: mateuslima90@gmail.com
    name: Mateus Lima Fonseca
    url: http://swagger.io
  description: APIs to manage client to access IDP Provider
  title: Client Manager
  version: 1.0.0
servers:
- description: Production Server
  url: http://localhost:8080/api/v1/idp
- description: UAT Server
  url: http://localhost:8081/api/v1/idp
- description: TST Server
  url: http://localhost:8082/api/v1/idp
- description: DEV Server
  url: http://localhost:8083/api/v1/idp
paths:
  /clients:
    get:
      description: Return all clients from IDP Provider
      operationId: getAllClient
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Client'
                type: array
          description: A JSON array of clients
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Not Found
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Return a list of clients
      tags:
      - clients
      x-accepts: application/json
  /client:
    post:
      description: Return a specific client
      operationId: createSingleClient
      requestBody:
        $ref: '#/components/requestBodies/inline_object'
        content:
          application/json:
            schema:
              properties:
                id:
                  example: ae1271te127ts178t27e1
                  type: string
                username:
                  example: integration-notification-api
                  type: string
                description:
                  example: Integration Notification API
                  type: string
                owner:
                  example: mateuslima90@gmail.com
                  type: string
                team:
                  example: APi Automation Team
                  type: string
              type: object
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
          description: A JSON array of clients
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Invalid Payload
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Client already exists
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a client
      tags:
      - clients
      x-contentType: application/json
      x-accepts: application/json
  /client/{clientId}:
    delete:
      description: Delete a specific client
      operationId: deleteSingleClient
      parameters:
      - description: ID of pet to fetch
        explode: false
        in: path
        name: clientId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
          description: A JSON array of clients
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete Client by Id
      tags:
      - clients
      x-accepts: application/json
    get:
      description: Return a specific client
      operationId: getSingleClient
      parameters:
      - description: ID of pet to fetch
        explode: false
        in: path
        name: clientId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
          description: A JSON array of clients
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Search Client by Id
      tags:
      - clients
      x-accepts: application/json
    put:
      description: Update a specific client
      operationId: updateSingleClient
      parameters:
      - description: ID of pet to fetch
        explode: false
        in: path
        name: clientId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/inline_object_1'
        content:
          application/json:
            schema:
              properties:
                username:
                  example: integration-notification-api
                  type: string
                description:
                  example: Integration Notification API
                  type: string
                owner:
                  example: mateuslima90@gmail.com
                  type: string
                team:
                  example: APi Automation Team
                  type: string
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
          description: A JSON array of clients
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Invalid Payload
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Client already exists
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Update Client by Id
      tags:
      - clients
      x-contentType: application/json
      x-accepts: application/json
components:
  requestBodies:
    inline_object_1:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object_1'
      required: true
    inline_object:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/inline_object'
      required: true
  schemas:
    Client:
      example:
        owner: mateuslima90@gmail.com
        description: User responsible to create bulk crm tasks
        id: 121dasda212e12dsawd3
        team: APi Automation Team
        username: agLAl0hgqNKwp9APtAW6
      properties:
        id:
          example: 121dasda212e12dsawd3
          type: string
        username:
          example: agLAl0hgqNKwp9APtAW6
          type: string
        description:
          example: User responsible to create bulk crm tasks
          type: string
        owner:
          example: mateuslima90@gmail.com
          type: string
        team:
          example: APi Automation Team
          type: string
      type: object
      xml:
        name: user
      x-swagger-router-model: io.swagger.petstore.model.User
    ApiResponse:
      properties:
        code:
          example: 404
          format: int64
          type: integer
        msg:
          example: agLAl0hgqNKwp9APtAW6
          type: string
      type: object
      xml:
        name: ApiResponse
      x-swagger-router-model: io.swagger.petstore.model.ApiResponse
    inline_object:
      properties:
        id:
          example: ae1271te127ts178t27e1
          type: string
        username:
          example: integration-notification-api
          type: string
        description:
          example: Integration Notification API
          type: string
        owner:
          example: mateuslima90@gmail.com
          type: string
        team:
          example: APi Automation Team
          type: string
      type: object
    inline_object_1:
      properties:
        username:
          example: integration-notification-api
          type: string
        description:
          example: Integration Notification API
          type: string
        owner:
          example: mateuslima90@gmail.com
          type: string
        team:
          example: APi Automation Team
          type: string
      type: object
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http